Updated to use custom version information
The custom version is cool, because it makes is easy to see what firmware is
running on my board. Obviously, this was useful to make sure that the code I
compiled was actually running (without having to modify some LED somewhere).

I noticed the '__has_include()' macro wasn't working right. Notice we check it
exists first (which it does), and then try to test the header
exists. Unfortunately, we're missing the '../../' when we test - so the compiler
can't find our header.

This change assumes the header is there if a custom version file was enabled. A
compiler error, in my opinion, is much more appropriate as silently ignoring when
I explicitly ask for a custom header is confusing!

Note the quick test this works after building:

  strings firmware.bin | grep dancollins # ... or whatever custom info you added